==> R/03_oldSNPFiltering.Rmd <==
---
title: "SNP Filtering"
author: "Steve Pederson"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE, warning = FALSE)
```

```{r}
library(tidyverse)
library(magrittr)
library(rtracklayer)
library(GenomicRanges)
library(AnnotationHub)
library(parallel)
library(reshape2)
library(pander)
library(scales)
library(stringr)
nCores <- min(12, detectCores() - 1)
```


# Introduction

This file takes the raw output from `Stacks` and filters the complete set of SNPs to arrive at the final list of SNPs used for downstream analysis.
The two populations are referred to below as Population `1` (*Gum Creek*, 1996) and Population `2` (*Oraparinna*, 2012).

## Outline of selection process for SNPs to be analysed

The filtering steps involved are:

1. Removal of SNPs located on sex chromosomes
2. Removal of SNPs within the *PstI* restriction site errors
3. Removal of *fixed* alleles
4. Removal of SNPs not observed in 1996 as these would be likely due to migration not selection
5. Removal of SNPs identified in <75% of both populations
6. Removal of SNPs within the same GBS tag, with identical allele frequencies
7. Removal of SNPs called by `Stacks` to be in separate tags, but which share identical genomic co-ordinates
8. Selection of SNPs within 40kb of a gene

`Stacks` identifies `P` (major) and `Q` (minor) alleles within each population, and an additional processing step was included to ensure the `P` allele in both populations was the SNP determined to be the `P` allele in the 1996 population.

# Data Setup

#### Genome Setup

Whilst aligned to the RefSeq/NCBI genome build, for compatibility with Ensembl gene models, a mapping object was generated from the [RefSeq Assembly Report](ftp://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Oryctolagus_cuniculus/latest_assembly_versions/GCF_000003625.3_OryCun2.0/GCF_000003625.3_OryCun2.0_assembly_report.txt)

```{r ncbi, cache=TRUE}
ncbiReport <- url("ftp://ftp.ncbi.nih.gov/genomes/refseq/vertebrate_mammalian/Oryctolagus_cuniculus/latest_assembly_versions/GCF_000003625.3_OryCun2.0/GCF_000003625.3_OryCun2.0_assembly_report.txt") %>%
    read_delim(delim = "\t", col_names = FALSE, na = "na", comment = "#") %>%
  set_colnames(c("Sequence-Name", "Sequence-Role", "Assigned-Molecule", "Assigned-Molecule-Location/Type", "GenBank-Accn", "Relationship", "RefSeq-Accn", "Assembly-Unit", "Sequence-Length", "UCSC-style-name")) %>%
  filter(!is.na(`RefSeq-Accn`))
ensUsesGB <- grepl("scaffold", ncbiReport$`Sequence-Role`)
ncbiReport$Ensembl <- ncbiReport$`Sequence-Name`
ncbiReport$Ensembl[ensUsesGB] <- gsub("\\.[0-9]$", "", ncbiReport$`GenBank-Accn`[ensUsesGB])
ncbi2Ens <- with(ncbiReport, data.frame(Chr = Ensembl, 
                                        Length = `Sequence-Length`,
                                        row.names = `RefSeq-Accn`))
```

Gene information was obtained from Ensembl Build 94 and loaded 

```{r ensGenes, cache=TRUE}
ah <- AnnotationHub() %>%
    subset(species == "Oryctolagus cuniculus") %>%
    subset(dataprovider == "Ensembl")
ensDb <- ah[["AH64956"]]
ensGenes <- genes(ensDb)
ensTwoBit <- ah[["AH65952"]]
ensSeqinfo <- seqinfo(ensTwoBit)
```

#### Loading SNP Data

```{r sumData}
sumData <- file.path("..", "5_stacks", "stacks", "populations.sumstats.tsv.gz") %>%
    gzfile() %>%
    read_delim(delim = "\t", col_names = TRUE, skip = 3,
               col_types = "iciiicciddddddddddddi") %>%
    set_colnames(str_replace_all(colnames(.), "# ", "")) %>%
    dplyr::select(-contains("Batch"), -contains("Pi"), -contains("Fis")) %>%
    mutate(snpID = paste(`Locus ID`, Col, sep="_"),
           Private = as.logical(Private)) %>%
    dplyr::filter(`Pop ID` != 3, Chr != "X") %>%
    dplyr::select(Chr, BP, `Locus ID`, Col, snpID, 
                  `Pop ID`, contains("Nuc"),  N, P, 
                  contains("Obs"), contains("Exp"), Private) %>%
    group_by(snpID) %>%
    dplyr::filter(n() == 2) %>%
    ungroup()
```

The `stacks` summary output was loaded immediately removing:

- SNPs on Chromosome X
- Data from the Turretfield outgroup, which was not required for this stage of the analysis

This object contained information for `r comma(length(unique(sumData$snpID)))` SNPs across `r comma(length(unique(sumData[["Locus ID"]])))` GBS tags.

#### Conversion of SNPs to GRanges

The genomic locations of all SNPs were then prepared as a `GRanges` object.

```{r snp2GR}
snp2GR <- sumData %>%
  distinct(snpID, .keep_all = TRUE) %>%
  makeGRangesFromDataFrame(keep.extra.columns=TRUE,
                           seqinfo = ensSeqinfo,
                           seqnames.field = "Chr",
                           start.field = "BP",
                           end.field = "BP",
                           ignore.strand = TRUE) %>%
  extract(, c("Locus ID", "snpID")) %>%
  set_names(.$snpID) %>%
  sort
```


# SNP Filtering And Selection

## Uninformative SNPs

- Fixed SNPs were defined as `P > 0.95` in both populations or `P = 0.5` in both populations

```{r}
fixedSNPs <- sumData %>%
  group_by(snpID) %>%
  summarise(minP = min(P),
            maxP = max(P),
            both50 = all(minP == 0.5, maxP == 0.5)) %>%
  dplyr::filter(minP > 0.95 | both50) %>%
  extract2("snpID")
```

- SNPs not detected in the 1996 population were identified as these were likely due to migration or some other factor, not selection

```{r}
uniq2Ora <- dplyr::filter(sumData, `Pop ID` == 1, P == 1)$snpID
```

- SNPs were also excluded if not identified in > 75% of both populations. However, this was performed by stacks using generation of haplotypes.

```{r}
minSamp <- sumData %>% 
  group_by(`Pop ID`) %>% 
  summarise(N = max(N)) %>%
  mutate(minN = ceiling(0.75*N)) %>%
  dplyr::select(`Pop ID`, N, minN)
```

```{r, echo=FALSE}
minSamp %>%
  dplyr::select(`Pop ID`, N, `Min Required` = minN) %>%
  pander(justify = "lrr",
         style = "rmarkdown",
         caption = "Minimum sample sizes required in each population.")
```

These filtering steps identified SNPs for exclusion as seen in the following table:

| Reason For Removal  | SNPs Marked For Removal       |
|:------------------- | -------------------------------:|
| Fixed Alleles       | `r comma(length(fixedSNPs))`    |
| Not Found in 1996   | `r comma(length(uniq2Ora))`     |
| **Total**           | **`r comma(length(unique(c(fixedSNPs, uniq2Ora))))`** |

Table: SNPs identified for removal

## Duplicate SNPs

```{r rmRedundant}
bestInTag <- sumData %>% 
    dplyr::filter(!snpID %in% c(fixedSNPs, uniq2Ora, lowSamp)) %>%
    dplyr::select(`Locus ID`, Col, snpID, `Pop ID`, N) %>%
    dcast(`Locus ID` + Col + snpID ~ `Pop ID`) %>% 
    distinct(`Locus ID`, `1`, `2`, .keep_all = TRUE) %>%
    extract2("snpID")
```

- Where multiple SNPs were present in a tag with identical allele frequencies and sample sizes, one was selected at random
- This marked `r comma(length(setdiff(setdiff(names(snp2GR), c(fixedSNPs, uniq2Ora, lowSamp)), bestInTag)))` SNPs as containing duplicated information

```{r besDuplicate}
bestDuplicate <- sumData %>% 
  dplyr::filter(snpID %in% bestInTag) %>%
  group_by(Chr, BP, snpID) %>%
  summarise(N = sum(N)) %>%
  ungroup() %>%
  mutate(`Locus ID` = gsub("([0-9]+)_[0-9]","\\1", snpID)) %>%
  arrange(Chr, BP, N, `Locus ID`) %>%
  distinct(Chr, BP, .keep_all = TRUE) %>%
  extract2("snpID")
```

- Some SNPs were identified at the same genomic location in two loci and these were assumed to be due to differing cut sites, or tags aligned to complementary strands
- If samples sizes were identical, otherwise the SNP with the largest total samples was chosen
- This gave `r comma(length(bestDuplicate))` candidate SNPs for downstream analysis.


## Selection of SNPs Near Genes

```{r finalSNPa}
finalSNPs <- snp2GR %>%
  subset(snpID %in% bestDuplicate) %>%
  resize(80001, fix = "center") %>%
  trim() %>%
  subsetByOverlaps(ensGenes) %>%
  names()
```

```{r nGenes}
nGenes <- snp2GR %>%
  extract(finalSNPs) %>%
  resize(80001, fix = "center") %>%
  trim() %>%
  findOverlaps(ensGenes) %>%
  subjectHits() %>%
  unique() %>%
  length()
```

Only SNPs within 40kb of a known gene were then retained, giving the **final total of `r comma(length(finalSNPs))` SNPs** for downstream analysis, which potentially provided markers for `r comma(nGenes)` genes.

## Genomic Coverage After SNP Filtering

```{r windowCounts}
winSize <- 1e05
windows <- tileGenome(seqinfo(snp2GR)[as.character(1:21),], tilewidth = 1e05)
windowCounts <- countOverlaps(windows, subset(snp2GR, snpID %in% finalSNPs)) 
```

Using `r as.integer(winSize/1000)`kb tiled windows across the autosomes, this final list of SNPs gave `r sum(windowCounts == 0)` windows without a SNP, and `r sum(windowCounts > 0)` windows with at least one SNP.
This means that `r percent(mean(windowCounts > 0))` of windows were sampled to a high enough quality using the GBS approach.
Of the windows with a SNP, the median number of SNPs in each window was `r median(windowCounts[windowCounts > 0])`, with the most densely sampled window containing `r max(windowCounts)` SNPs.
This is in comparison to the dataset before filtering of low-quality/duplicate SNPs, in which `r percent(mean(countOverlaps(windows, snp2GR) > 0))` of the `r as.integer(winSize/1000)`kb windows contained at least one SNP.


# Correction of P & Q Nucleotides

P and Q alleles are defined separately for each population, and need to be redefined such that these alleles are common across populations.
Define each SNP such that the P (i.e. most common) allele in Population 1 is considered as the reference allele.

```{r reDefPQ}
sumData %<>%
  dplyr::filter(snpID %in% finalSNPs) %>%
  split(f = .$snpID) %>%
  mclapply(function(x){
    if (length(unique(x$`P Nuc`)) == 2){ # If there are 2 P alleles
      x[2, c("P Nuc", "Q Nuc")] <- x[2, c("Q Nuc", "P Nuc")]
      x$P[2] <- 1 - x$P[2]
    }
    x
  }, mc.cores = nCores) %>%
  bind_rows
```


```{r outFile}
gzOut <- file.path("..", "5_stacks", "stacks", "filteredSNPs.tsv.gz") %>%
  gzfile("w")
sumData %>%
  dplyr::select(-Private) %>%
  write_tsv(gzOut)
close(gzOut)
```

This final object was exported as the object `filteredSNPs.tsv.gz`


# Editing of genpop data

The `genpop` file as output by `stacks` was loaded into `R` and converted to genotypes manually.


- First extract all the setup information

```{r}
gpFile <- file.path("..", "5_stacks", "genepop", "populations.snps.genepop.gz") %>% gzfile()
gpData <- readLines(gpFile)
close(gpFile)
popStart <- grep("^pop", gpData) + 1
popEnd <- c(popStart[-1] - 2, length(gpData))
nPops <- length(popStart)
sampleRows <- seq_along(popStart) %>%
  sapply(function(x){seq(popStart[x], popEnd[x])}) %>%
  unlist()
sampleIDs <- gsub("(gc|ora|TF|tf|Y|pt)([0-9ABC]+),.+", "\\1\\2", gpData[sampleRows])
snpIDs <- gpData[2] %>% str_split(",") %>% extract2(1)
```

- Now get the alleles using the existing allele encoding
- Then restrict to the SNPs being analysed

---
# Failure from Here
---

```{r}
alleles <- gsub("(gc|ora|TF|tf|Y|pt)[0-9ABC]+,\\t", "", gpData[sampleRows]) %>%
  str_split_fixed("\t", n = length(snpIDs)) %>%
  replace(. == "0000", NA) %>%
  set_colnames(snpIDs) %>%
  set_rownames(sampleIDs) %>%
  magrittr::extract(,colnames(.) %in% finalSNPs)
```

- Convert to minor allele counts

```{r}
minorAlleleCounts <- alleles %>%
  apply(MARGIN = 2, function(x){
    allNums <- sort(unique(x[!is.na(x)])) # The existing allele combinations
    allIDs <- unique(c(substr(allNums, 1, 2),
                       substr(allNums, 3, 4))) # The existing allele numbers
    # Create all possible genotypes
    gTypes <- c(paste0(allIDs[1], allIDs[1]),
                  paste0(allIDs[1], allIDs[2]),
                  paste0(allIDs[2], allIDs[2]))
    # Find the maximum allele based on homozygotes across all populations
    homs <- vapply(gTypes[c(1, 3)],function(y){sum(grepl(y, x), na.rm = TRUE)}, integer(1))
    P <- which.max(homs)
    # Reverse the genotypes & use as factor levels
    if (P == 2) gTypes <- rev(gTypes)
    as.integer(factor(x, levels = gTypes)) - 1
  }) %>%
  set_rownames(sampleIDs)
```

```{r}
gzOut <- file.path("..", "data", "minorAlleleCounts.tsv.gz") %>%
  gzfile("w")
minorAlleleCounts %>%
  as.data.frame() %>%
  rownames_to_column("sampleID") %>%
  write_tsv(gzOut)
close(gzOut)
```

This file was then exported as `minorAlleleCounts.tsv.gz`.
An additional genepop file was created for use with Bayescan and for estimation of effective population sizes.

```{r, results='hide'}
customGenepopFile <- file.path("..", "data", "filteredSNPs.genepop.gz") 
# The header
write_lines(x = "Created by custom script; Genepop version 4.1.3; Nov 05, 2017", 
           path = gzfile(customGenepopFile))
# Allele names
colnames(alleles) %>% 
  paste(collapse = ",") %>%
  write_lines(gzfile(customGenepopFile), append = TRUE)
# The first population
write_lines("pop", gzfile(customGenepopFile), append = TRUE)
# Allele data for each gc sample
grep("gc", rownames(alleles), value = TRUE) %>%
  lapply(function(x){
    alls <- replace(alleles[x,], is.na(alleles[x,]), "0000") %>%
      paste(collapse = "\t")
    paste(x, alls, sep = ",") %>%
      write_lines(gzfile(customGenepopFile), append = TRUE)
  })
# The next population
write_lines("pop", gzfile(customGenepopFile), append = TRUE)
# Allele data for each ora sample
grep("ora", rownames(alleles), value = TRUE) %>%
  lapply(function(x){
    alls <- replace(alleles[x,], is.na(alleles[x,]), "0000") %>%
      paste(collapse = "\t")
    paste(x, alls, sep = ",") %>%
      write_lines(gzfile(customGenepopFile), append = TRUE)
  })
```


```{r, echo=FALSE, results='asis'}
pander(sessionInfo()) 
```
