==> R/03_SNPFiltering.Rmd <==
---
title: "SNP Filtering"
author: "Steve Pederson"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    fig_caption: yes
    fig_height: 6
    fig_width: 10
    toc: true
    toc_float: true
    code_folding: hide   
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  warning = FALSE, message = FALSE
)
```

## Setup

```{r loadPackages}
library(SeqArray)
library(SNPRelate)
library(pander)
library(scales)
library(magrittr)
library(tidyverse)
theme_set(theme_bw())
if (interactive()) here::here("R") %>% setwd
```

This workflow goes through multiple steps involved in processing the stacks output:

1. Beginning with the VCF, this will be converted to a GDS object for easier interaction.
As this is a slow conversion process, this will only be performed once.
The resulting object will be used to prune SNPs in Linkage Disequilibrium.
Notably, this VCF is not available in this github repo as it is ~92Mb and too large.
Likewise the prepared GDS object is 86.6Mb, which is also too large for storage on github.


### VCF Conversion to GDS

```{r vcf2gds}
gdsPath <- file.path("..", "5_stacks", "gds", "populations.snps.gds")
makeGds <- !file.exists(gdsPath)
if (makeGds) {
    dir.create(dirname(gdsPath))
    file.path("..", "5_stacks", "vcf", "populations.snps.vcf.gz") %>%
        seqVCF2GDS(gdsPath, reference = "OryCun2.0")
}
gdsFile <- seqOpen(gdsPath, readonly = FALSE)
```

Sample information and subsequent population information was then extracted and added to the GDS file.

```{r sampleID}
sampleID <- seqGetData(gdsFile, "sample.id") %>%
    as.data.frame(stringsAsFactors = FALSE) %>%
    set_names("Sample") %>%
    mutate(Population = case_when(
        grepl("gc", Sample) ~ 1996L,
        grepl("ora", Sample) ~ 2012L,
        !grepl("(gc|ora)", Sample) ~ 2010L
    ),
    Location = case_when(
        Population == 1996 ~ "Gum Creek",
        Population == 2012 ~ "Oraparinna",
        Population == 2010 ~ "Turretfield"
    )) %>%
    as_tibble()
```

```{r addAnnotation}
if (makeGds) {
  add.gdsn(gdsFile, "sample.annotation", sampleID, replace = TRUE)
  seqClose(gdsFile)
  gdsFile <- seqOpen(gdsPath, readonly = FALSE)
}
```

### SNP Data Summary

```{r chromosomes}
chromosomes <- paste0("chr", c(1:21, "X"))
scaffolds <- seqGetData(gdsFile, "chromosome") %>%
    unique() %>%
    setdiff(chromosomes)
```

```{r snpSummary}
snpSummary <- tibble(
    variant.id = seqGetData(gdsFile, "variant.id") ,
    chromosome = seqGetData(gdsFile, "chromosome"),
    position = seqGetData(gdsFile, "position")
) %>%
    cbind(seqGetData(gdsFile, "annotation/format/DP")$data %>% 
              t %>%
              set_colnames(sampleID$Sample)) %>%
    as_tibble()
mnDist <- snpSummary %>% 
  dplyr::filter(chromosome %in% 1:21) %>% 
  split(f = .$chromosome) %>% 
  lapply(function(x){diff(x$position)}) %>% 
  unlist() %>% 
  mean() %>% 
  divide_by(1000) %>% 
  round(2)
mnDiffLoc <- snpSummary %>% 
  dplyr::filter(chromosome %in% 1:21) %>% 
  split(f = .$chromosome) %>%
  lapply(function(x){diff(x$position) %>% .[. > 100]}) %>%
  unlist() %>% 
  mean() %>% 
  divide_by(1000) %>% 
  round(2)
```

From this initial summary, `r comma(nrow(snpSummary))` unique SNPs were identified by the Stacks pipeline.
The mean distance between all autosomal SNPs was calculated to be `r mnDist`kb, however, many of these were within the same locus.
Excluding SNPs within the same locus (i.e. <100nt apart), the mean distance between SNPs became `r mnDiffLoc`kb.

Checking the call rate across all samples, 5x1996 samples with SNP call rates < 50% were identified. 
This value was not exceeded by any samples in the 2010 or 2012 populations.
No SNPs were found being unique to Turretfield, which is as expected by the settings given to `stacks`

```{r plotCallRate, fig.cap="*Summary of SNP call rate in all samples*", fig.height=7}
snpSummary %>%
    mutate_at(sampleID$Sample, list(is.na)) %>%
    summarise_at(sampleID$Sample, list(sum)) %>%
    gather(key = "Sample", value = "Missing") %>%
    mutate(Called = nrow(snpSummary) - Missing) %>%
    gather(key = "Type", value = "SNPs", -Sample) %>%
    left_join(sampleID) %>%
    ggplot(aes(Sample, SNPs / nrow(snpSummary), fill = Type)) +
    geom_bar(stat = "identity") +
    geom_hline(yintercept = 0.5, linetype = 2) +
    facet_wrap(~Population, scales = "free") +
    coord_flip() +
    labs(y = "Percent Called") +
    scale_y_continuous(expand = expand_scale(0, 0), labels = percent) +
    scale_fill_manual(values = c("green", "red"))
```

### LD Pruning

```{r ldPruning, results='hide'}
set.seed(1523)
minMaf <- 0.05
missingRate <- 0.25
minCor <- 0.4
snpset <- gdsFile %>%
    snpgdsLDpruning(
        sample.id = grep("(gc|ora)", seqGetData(., "sample.id"), value = TRUE),
        autosome.only = FALSE,
        maf = minMaf,
        missing.rate = missingRate,
        method = "corr",
        ld.threshold = minCor
        ) %>%
    .[setdiff(names(.), "chrX")]
seqResetFilter(gdsFile) 
```

SNPs were pruned for Linkage Disequilibrium using only samples from the 1996 and 2012 populations and the following criteria:

- Minor Allele Frequency > `r percent(minMaf)`
- Missing Data in < `r percent(missingRate)` of samples
- Correlation of > `r percent(minCor)` indicating LD


```{r keepSNPs}
keepSNPs <- c("variant.id", "position", "chromosome") %>%
    sapply(function(x){
        seqGetData(gdsFile, x)
    }, simplify = FALSE) %>%
    as_tibble() %>%
    dplyr::filter(variant.id %in% unlist(snpset))
```

```{r plotKeepSNPs, fig.cap = "*Summary of SNPs per chromosome after filtering for LD.*"}
keepSNPs %>%
    dplyr::filter(chromosome %in% c(1:22, "X")) %>%
    group_by(chromosome) %>%
    tally() %>%
    mutate(chromosome = factor(chromosome, levels = 1:21)) %>%
    droplevels() %>%
    arrange(chromosome) %>%
    ggplot(aes(chromosome, n)) +
    geom_bar(stat = "identity") +
    labs(x = "Primary Chromosomes",
         y = "Number of SNPs retained for initial analysis") +
    scale_x_discrete()
```


This set of SNPs was then exported as an R object for downstream use.

```{r}
write_rds(keepSNPs, "keepSNPsAfterLDPruning.RDS")
```

## SessionInfo

```{r, echo=FALSE}
pander(sessionInfo())
```
